+++ src/main/kotlin/com/wafflestudio/csereal/core/academics/dto/CourseDto.kt
@@ -2,7 +2,6 @@ package com.wafflestudio.csereal.core.academics.dto
 
 import com.wafflestudio.csereal.common.enums.LanguageType
 import com.wafflestudio.csereal.core.academics.database.CourseEntity
-import com.wafflestudio.csereal.core.resource.attachment.dto.AttachmentResponse
 
 data class CourseDto(
     val id: Long,
